+++ anni-repo/src/models/mod.rs
@@ -7,3 +7,8 @@ pub use album::*;
 pub use date::*;
 pub use repo::*;
 pub use tag::*;
+
+#[cfg(feature = "json")]
+mod json;
+#[cfg(feature = "json")]
+pub use json::*;